fastapi_et/requirements.txt
fastapi
uvicorn
slowapi
httpx
python-dotenv
